refactor: Moved option file into Common package
