+++ src/main/java/edu/lorsenmarek/backend/common/PageOptions.java
@@ -1,4 +1,4 @@
-package edu.lorsenmarek.backend.utility;
+package edu.lorsenmarek.backend.common;
 
 import lombok.*;
